diabetes-api/requirements.txt

uvicorn>=0.20.0,<0.30.0
fastapi>=0.88.0,<1.0.0
python-multipart>=0.0.5,<0.1.0
pydantic>=1.10.4,<1.12.0
typing_extensions>=4.2.0,<5.0.0
loguru>=0.5.3,<1.0.0
# fetched from gemfury
diabetes-regression-model==0.0.4 --extra-index-url=${PIP_EXTRA_INDEX_URL}
